Update package.json from utils package
Removed type module from package.json to fix: 

Module not found: Error: Can't resolve './components' in 'PROJECT_PATH/node_modules/@react-slip-and-slide/utils/dist'
Did you mean 'index.js'?
BREAKING CHANGE: The request './components' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.