feat(EngineMode): create a clear distinction between rendering modes based on the prop `infinite` (#3)
Now, items will only self translate if the props `infinite` or
`visibleItems` are defined. Otherwise the translation will happen on the
element wrapping all the items. This besides being conceptually more
correct, it also improves the performance and allows the items to follow
the flexbox flow by default, meaning that the item measurement is not
required for defining it's displacement.